refactor(src): move vertical glue under src/verticals with root compat
