+++ src/verticals/minecraft/query_live_action.rs
