+++ src/verticals/minecraft/session.rs
